+++ backend/changelog.md
@@ -1,14 +1,24 @@
-v0.2 :
+## v0.3 :
+
+- create and save user session
+- better auth middleware handling
+- better and reusable render function
+- added config file for project
+- git ignore dataStorage folder
+- some refactoring
+
+## v0.2 :
+
 - httponly cookie
 - SSR for profile page
 - logout (now available on profile page)
 - password hashing
 - store user id instead of email in cookie
 
-v0.1 :
-- Login & signup page 
+## v0.1 :
+
+- Login & signup page
 - authentication with cookie
 - dynamic navigation
 - profile page
-- some refactoring 
-
+- some refactoring